+++ CHANGELOG.md
@@ -163,3 +163,4 @@
 - 2023-11-08: Pool find email, ready pool, Makefile, main, +3
 - 2023-11-09: Init frontier, discover, leads, deal, +6
 - 2023-11-10: Pagination email opener, descend, users, follow up, +2
+- 2023-11-10: Cache descend, sender, Makefile, leads, +4